# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HSE Mobile Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: List templates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplateById
      tags: [templates]
      summary: Get template with questions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateWithQuestions"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        questionCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - questionCount
        - createdAt
    TemplateWithQuestions:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
      required:
        - id
        - name
        - questions
        - createdAt
    CreateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    Question:
      type: object
      properties:
        id:
          type: integer
        templateId:
          type: integer
        text:
          type: string
        answerType:
          type: string
          enum: [yes_no, text]
        isMandatory:
          type: boolean
        requiresPhoto:
          type: boolean
        categoryId:
          type: integer
        categoryName:
          type: string
        orderIndex:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - templateId
        - text
        - answerType
        - isMandatory
        - requiresPhoto
        - orderIndex
        - createdAt
...